Workout and completedWorkout Stats initial implementation
Get workout volume per muscle group and get weekly volume (per muscle group) service/workoutsStats.service that calculates the stats from the database. stats.controller to get them in the frontend or it is also possible to get the workout stats directly from the workout/completeworkout model in the database because they are saved there.